+++ src/lessons.ts
@@ -1,6 +1,12 @@
 import * as lesson01 from "./lesson-01";
+import * as lesson02 from "./lesson-02";
+import * as lesson03 from "./lesson-03";
+import * as lesson04 from "./lesson-04";
 
 export const LESSONS = {
   "01": lesson01,
+  "02": lesson02,
+  "03": lesson03,
+  "04": lesson04,
 } as const;
 export type LessonId = keyof typeof LESSONS;